requirements.txt
arabic-reshaper==2.1.3
asgiref==3.3.4
cycler==0.10.0
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.4
django-crispy-forms==1.11.2
django-import-export==2.5.0
django-pdfkit==0.3.1
django-uuslug==1.2.0
django-widget-tweaks==1.4.8
django-wkhtmltopdf==3.3.0
et-xmlfile==1.1.0
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
kiwisolver==1.3.1
MarkupPy==1.14
matplotlib==3.3.4
numpy==1.19.5
odfpy==1.4.1
openpyxl==3.0.7
pdfkit==0.6.0
Pillow==8.2.0
psycopg2==2.8.6
pyparsing==2.4.7
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
python-slugify==5.0.2
pytz==2021.1
PyYAML==5.4.1
reportlab==3.5.67
scipy==1.5.4
six==1.16.0
sqlparse==0.4.1
static3==0.7.0
tablib==3.0.0
text-unidecode==1.3
typing-extensions==3.10.0.0
webencodings==0.5.1
xhtml2pdf==0.2.5
xlrd==2.0.1
xlwt==1.3.0
